解决ubuntu上执行部署脚本失败的问题
