--- src/main/groovy/com/sankuai/srq/deploy/script/Hello.java
@@ -1,7 +1,0 @@
-package com.sankuai.srq.deploy.script;
-
-public class Hello {
-	public static void main(String[] args) {
-		System.out.println("Hello,java");
-	}
-}
